requirements.txt
selenium
beautifulsoup4
python-dotenv
requests
